+++ R/arrange.R
@@ -21,7 +21,7 @@
 #'   group_by(x1) |>
 #'   arrange(-x2, .by_group = TRUE)
 
-arrange.DataFrame <- function(.data, ..., .by_group = FALSE) {
+arrange.RPolarsDataFrame <- function(.data, ..., .by_group = FALSE) {
 
   check_polars_data(.data)
 
@@ -65,4 +65,4 @@ arrange.DataFrame <- function(.data, ..., .by_group = FALSE) {
 }
 
 #' @export
-arrange.LazyFrame <- arrange.DataFrame
+arrange.RPolarsLazyFrame <- arrange.RPolarsDataFrame