+++ R/pull.R
@@ -4,7 +4,7 @@
 #'
 #' @param .data A Polars Data/LazyFrame
 #' @param var A quoted or unquoted variable name
-#' @inheritParams slice_tail.DataFrame
+#' @inheritParams slice_tail.RPolarsDataFrame
 #'
 #' @export
 #' @examplesIf require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)
@@ -12,17 +12,17 @@
 #' pull(pl_test, Sepal.Length)
 #' pull(pl_test, "Sepal.Length")
 
-pull.DataFrame <- function(.data, var, ...) {
+pull.RPolarsDataFrame <- function(.data, var, ...) {
   check_polars_data(.data)
   var <- tidyselect_named_arg(.data, rlang::enquo(var))
   # for testing only
-  if (inherits(.data, "LazyFrame") && Sys.getenv("TIDYPOLARS_TEST") == "TRUE") {
+  if (inherits(.data, "RPolarsLazyFrame") && Sys.getenv("TIDYPOLARS_TEST") == "TRUE") {
     return(to_r(.data$collect()$select(pl$col(var)))[[1]])
   }
 
   to_r(.data$select(pl$col(var)))[[1]]
 }
 
-#' @rdname pull.DataFrame
+#' @rdname pull.RPolarsDataFrame
 #' @export
-pull.LazyFrame <- pull.DataFrame
+pull.RPolarsLazyFrame <- pull.RPolarsDataFrame