+++ R/utils-expr.R
@@ -216,7 +216,7 @@ translate_expr <- function(.data, quo, new_vars = NULL, env) {
             suppressWarnings({
               tr <- try(do.call(fn, list(args)), silent = TRUE)
             })
-            if (inherits(tr, "Expr")) {
+            if (inherits(tr, "RPolarsExpr")) {
               return(tr)
             } else {
               abort(
@@ -324,6 +324,16 @@ clean_dots <- function(...) {
   dots <- get_dots(...)
   dots[["__tidypolars__new_vars"]] <- NULL
   dots[["__tidypolars__env"]] <- NULL
+  caller_call <- deparse(rlang::caller_call()[[1]])
+  called_from_pl_paste <- length(caller_call) == 1 && caller_call %in% c("pl_paste", "pl_paste0")
+  dots <- lapply(dots, function(x) {
+    if (called_from_pl_paste &&
+        inherits(x, c("character", "logical", "double", "integer", "complex"))) {
+      pl$lit(x)
+    } else {
+      x
+    }
+  })
   dots
 }